feat(app): Add builder config + fix corpus selector

+++ juxtorpus/corpus/app.py
@@ -42,7 +42,7 @@ class App(object):
 
         # widgets
         self._corpus_selector = None
-        self._corpus_builder = None
+        self._corpus_builder_configs = None
 
     ## Corpus Registry ##
     def update_registry(self, corpus_id, corpus):
@@ -96,9 +96,7 @@ class App(object):
 
         def on_click_confirm(_):
             selected_files = [d.get('path') for d in self._files.values() if d.get('selected')]
-            if len(selected_files) < 0:
-                print("No files selected.")
-                return
+            if len(selected_files) <= 0: return
             self._builder = CorpusBuilder(selected_files)
             # hbox_corpus_builder.children = tuple((Label(p.name) for p in self._builder.paths))
             hbox_corpus_builder = self._create_corpus_builder()
@@ -108,32 +106,85 @@ class App(object):
         return vbox
 
     # Widget: Corpus Builder
+
+    def _create_text_meta_dtype_row(self, id_: str, text_checked: bool, meta_checked: bool, config: dict):
+        dtypes = sorted(list(CorpusBuilder.allowed_dtypes))
+
+        label = widgets.Label(f"{id_}", layout=widgets.Layout(width='30%'))
+        t_checkbox = widgets.Checkbox(value=text_checked, layout=widgets.Layout(width='15%'))
+        t_checkbox.style.description_width = '0px'
+
+        m_checkbox = widgets.Checkbox(value=meta_checked, layout=widgets.Layout(width='15%'))
+        m_checkbox.style.description_width = '0px'
+
+        dtype_dd = widgets.Dropdown(options=dtypes, value=dtypes[0], disabled=False,
+                                    layout=widgets.Layout(width='100px'))
+
+        # dtype_dd.observe # todo
+        def _toggle_checkbox(event):
+            if not event.get('new'):
+                config['text'] = False
+                config['meta'] = False
+            else:
+                if id(event.get('owner')) == id(t_checkbox):
+                    m_checkbox.value = False
+                    dtype_dd.value = dtypes[dtypes.index('str')]  # if ticked as text - sets dtype as str
+                    config['text'] = True
+                    config['meta'] = False
+                elif id(event.get('owner')) == id(m_checkbox):
+                    t_checkbox.value = False
+                    dtype_dd.value = dtypes[dtypes.index('auto')]  # if ticked as text - sets dtype as str
+                    config['meta'] = True
+                    config['text'] = False
+
+        def _update_dtype(event):
+            config['dtype'] = event.get('new')
+
+        t_checkbox.observe(_toggle_checkbox, names='value')
+        m_checkbox.observe(_toggle_checkbox, names='value')  # todo: toggle text and meta
+        dtype_dd.observe(_update_dtype, names='value')
+        return widgets.HBox([label, t_checkbox, m_checkbox, dtype_dd])
+
     def _create_corpus_builder(self):
-        if self._builder is None: return VBox()
+        if self._builder is None: return VBox()  # self._builder must first be set up before.
 
-        def create_text_meta_dtype_row(id_: str, text_checked: bool, meta_checked: bool, dtypes: list[str]):
-            label = widgets.Label(f"{id_}", layout=widgets.Layout(width='30%'))
-            t_checkbox = widgets.Checkbox(value=text_checked, layout=widgets.Layout(width='15%'))
-            t_checkbox.style.description_width = '0px'
-            m_checkbox = widgets.Checkbox(value=meta_checked, layout=widgets.Layout(width='15%'))
-            m_checkbox.style.description_width = '0px'
-            m_checkbox.observe(lambda e: print(e), names='value')       # todo: toggle text and meta
-            dtype_dd = widgets.Dropdown(options=dtypes, value=dtypes[0], disabled=False,
-                                        layout=widgets.Layout(width='100px'))
-            return widgets.HBox([label, t_checkbox, m_checkbox, dtype_dd])
-
+        # creates the top labels.
         top_labels = [('id', '30%'), ('text', '15%'), ('meta', '15%'), ('dtype', '30%')]
         selection_top_labels = widgets.HBox(
             list(map(lambda ls: widgets.HTML(f"<b>{ls[0]}</b>", layout=widgets.Layout(width=ls[1])), top_labels)))
 
-        dtypes = sorted(list(CorpusBuilder.allowed_dtypes))
+        # creates all the rows
         columns = self._builder.columns
+        self._corpus_builder_configs = {col: {'text': False, 'meta': True} for col in columns}
+        configs = self._corpus_builder_configs
         selection_widgets = [selection_top_labels]
-        selection_widgets.extend([create_text_meta_dtype_row(col, False, True, dtypes) for col in columns])
-        # return HBox([Label("hello")], layout=Layout(width='100%', height='100%'))
-        return VBox(selection_widgets, layout=Layout(width='100%'))
+        gen_rows = (self._create_text_meta_dtype_row(col, check.get('text'), check.get('meta'), check)
+                    for col, check in configs.items())
+        selection_widgets.extend(gen_rows)
+
+        # create build button
+        key_textbox = widgets.Text(description='Corpus ID:', placeholder='ID to be stored in the registry')
+        corpus_id = {'name': ''}
+        key_textbox.observe(lambda event: corpus_id.update({'name': event.get('new')}), names='value')
+
+        button = Button(description='build')
+
+        def _on_click_build_corpus(_):
+            for key, config in configs.items():
+                if config.get('text'):
+                    self._builder.set_text_column(key)
+                else:
+                    dtype = config.get('dtype')
+                    self._builder.add_metas(key, dtype)
+            self.update_registry(corpus_id.get('name'), self._builder.build())
+
+        button.on_click(_on_click_build_corpus)
+
+        return HBox([VBox(selection_widgets, layout=Layout(width='70%')),
+                     VBox([key_textbox, button], layout=Layout(width='30%'))])
 
     ## Widget: Corpus Selector ##
+
     def corpus_selector(self):
         if self._corpus_selector is None:
             self._corpus_selector = GridBox([self._create_corpus_selector_table()],
@@ -151,9 +202,9 @@ class App(object):
         return VBox([hbox_registry_labels] + [self._create_corpus_selector_row(k) for k in self.REGISTRY.keys()])
 
     def _create_corpus_selector_row(self, corpus_id):
-        checkbox = widgets.Checkbox(description=f"{corpus_id}")
+        checkbox = widgets.Checkbox(description=f"{corpus_id}", layout=_create_layout(**corpus_id_layout))
         checkbox.style = {'description_width': '0px'}
-        checkbox.observe(self._observe_corpus_selector_checkbox)
+        checkbox.observe(self._observe_corpus_selector_checkbox, names='value')
 
         corpus = self.REGISTRY.get(corpus_id, None)
         if corpus is None: raise KeyError(f"Corpus ID: {corpus_id} does not exist.")
@@ -161,7 +212,6 @@ class App(object):
         parent_corpus = corpus.find_root()
         parent = list(self.REGISTRY.keys())[list(self.REGISTRY.values()).index(parent_corpus)]
         if parent == corpus_id: parent = ''
-        checkbox.layout = _create_layout(**corpus_id_layout)
         checkbox.add_class('corpus_id_focus_colour')  # todo: add this HTML to code
         return HBox([checkbox,
                      Label(str(size), layout=_create_layout(**size_layout)),
@@ -169,15 +219,17 @@ class App(object):
                     layout=hbox_layout)
 
     def _observe_corpus_selector_checkbox(self, event):
+        if self._corpus_selector is None: return
         value, owner = event.get('new'), event.get('owner')
         if value:
-            # deselect everything else
-            for child in self.corpus_selector():
-                for cb in child.children:
-                    if type(cb) == Checkbox:
-                        cb.value = False if cb != owner else True
             self._selected_corpus = self.REGISTRY.get(owner.description, None)  # todo: normalise this, then rm err
             if self._selected_corpus is None: raise RuntimeError("This should not happen.")
+            # deselect everything else
+            for vboxes in self._corpus_selector.children:
+                for hboxes in vboxes.children:
+                    for cb in hboxes.children:
+                        if type(cb) == Checkbox:
+                            cb.value = False if cb != owner else True
 
     ## Widget: Corpus Slicer ##
     def corpus_slicer(self):
